chore(FE): separate out layout managing and drawing
